Don't limit the upper bound of requests
Unless there's a very good reason to put a limit on the upper bound of the
dependencies in a library, it's good practice to not include those. We're using
a higher version of requests because we have another dependency that relies on
new requests>=2.9 features, so now there's an unresolvable conflict.