Stringify userId and anonymousId
Ref: https://segment.atlassian.net/browse/LIB-27

Our API truncates bignum values. This is problematic because this library has historically accepted Numbers as userIds and anonymousIds. TTo workaround the truncation, this library needs to stringify userId and anonymousId on the client.

The test cases are numbers that will be truncated in Node, e.g. `node -e "console.log(157963456373623802 + 1)"`.