Use semver-compatibility operator everywhere
This will prevent the library from accidentally installing a possibly
backwards-incompatible version.

For `requests`, the check is equivalent to >=2.7,<3.0
For `six` and `monotonic` it adds a constraint to prevent major version bumps
For `python-dateutil`, the check isn't theoritcally equivalent,
but in practice the only release after 2.1 is 2.2 so it should work the same.